[AN] feat: 더미데이터로 Rv 데이터 테스트 #629
