+++ android/app/src/main/java/com/onair/hearit/data/api/HearitService.kt
@@ -2,7 +2,6 @@ package com.onair.hearit.data.api
 
 import com.onair.hearit.data.dto.GroupedCategoryHearitResponse
 import com.onair.hearit.data.dto.HearitResponse
-import com.onair.hearit.data.dto.PlayingBookmarkResponse
 import com.onair.hearit.data.dto.RandomHearitResponse
 import com.onair.hearit.data.dto.RecentUploadResponse
 import com.onair.hearit.data.dto.RecommendHearitResponse
@@ -21,9 +20,6 @@ interface HearitService {
 
     @GET("api/v1/hearits/recent")
     suspend fun getRecentUploadHearits(): Response<List<RecentUploadResponse>>
-
-    //    @GET("api/v1/hearits/playing-bookmarks")
-    suspend fun getPlayingBookmarkHearits(): Response<List<PlayingBookmarkResponse>>
 
     @GET("api/v2/hearits/explore")
     suspend fun getRandomHearits(
