+++ android/app/src/main/java/com/onair/hearit/domain/repository/BookmarkRepository.kt
@@ -2,12 +2,15 @@ package com.onair.hearit.domain.repository
 
 import com.onair.hearit.domain.model.Bookmark
 import com.onair.hearit.domain.model.PageResult
+import com.onair.hearit.domain.model.PlayingBookmarkHearit
 
 interface BookmarkRepository {
     suspend fun getBookmarks(
         page: Int?,
         size: Int?,
     ): Result<PageResult<Bookmark>>
+
+    suspend fun getPlayingBookmarkHearits(): Result<List<PlayingBookmarkHearit>>
 
     suspend fun addBookmark(hearitId: Long): Result<Long>
