+++ android/app/src/main/java/com/onair/hearit/presentation/home/HomeViewModelFactory.kt
@@ -7,10 +7,12 @@ import com.onair.hearit.di.RepositoryProvider
 @Suppress("UNCHECKED_CAST")
 class HomeViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        val bookmarkRepository = RepositoryProvider.bookmarkRepository
         val hearitRepository = RepositoryProvider.hearitRepository
         val memberRepository = RepositoryProvider.memberRepository
         val playingHistoryRepository = RepositoryProvider.playingHistoryRepository
         return HomeViewModel(
+            bookmarkRepository,
             hearitRepository,
             memberRepository,
             playingHistoryRepository,
